fix(procmgr): apply legacy SCM env on privileged Windows spawns
Command::spawn only applied CreateEnvironmentBlock baseline, so
datadog-agent-process never inherited registry Environment from the
legacy datadog-process-agent SCM service.
