+++ pkg/procmgr/rust/src/spawn/request.rs
@@ -14,6 +14,7 @@ use crate::env::{expand_env_vars, parse_environment_file, try_expand_env_vars};
 use super::stdio::{StdioSetting, parse_stdio_setting, to_command_stdio};
 
 pub(crate) struct SpawnRequest {
+    process_name: String,
     command: String,
     args: Vec<String>,
     env: Vec<(String, String)>,
@@ -55,6 +56,7 @@ impl SpawnRequest {
 
     pub(crate) fn from_config(process_name: &str, config: &ProcessConfig) -> Result<Self> {
         Ok(Self {
+            process_name: process_name.to_string(),
             command: expand_env_vars(&config.command),
             args: config.args.iter().map(|a| expand_env_vars(a)).collect(),
             env: collect_env(process_name, config)?,
@@ -72,7 +74,10 @@ impl SpawnRequest {
         cmd.args(&self.args);
         cmd.env_clear();
         #[cfg(windows)]
-        crate::platform::apply_child_baseline_env(&mut cmd);
+        {
+            crate::platform::apply_child_baseline_env(&mut cmd);
+            crate::platform::apply_legacy_scm_env(&mut cmd, &self.process_name);
+        }
         for (k, v) in &self.env {
             cmd.env(k, v);
         }
